Updated to use GCDAsyncSocket in Podfile instead of directly in code Other changes

+++ source/DriveWireDocument.h
@@ -20,7 +20,7 @@
     IBOutlet NSDrawer *debugDrawer;
     IBOutlet NSButton *loggingSwitch;
     IBOutlet NSButton *statSwitch;
-	int lastPortSelected;
+	NSInteger lastPortSelected;
     NSWindowController *myWindowController;
 }
